requirements.txt
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
discord.py==1.2.3
firebase-admin==2.17.0
google-api-core==1.14.2
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.3.0
google-cloud-storage==1.17.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.22.0
httplib2==0.13.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
msgpack==0.6.1
multidict==4.5.2
protobuf==3.9.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pylint==2.3.1
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
